requirements.txt
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
cryptography==45.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
paramiko==3.5.1
pillow==11.2.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
python-telegram-bot==22.1
PyYAML==6.0.2
qrcode==8.2
sniffio==1.3.1
typing_extensions==4.13.2
